+++ unstructured/partition/html.py
@@ -2,6 +2,7 @@ from typing import IO, TYPE_CHECKING, Dict, List, Optional
 
 import requests
 
+from unstructured.chunking.title import add_chunking_strategy
 from unstructured.documents.elements import Element, process_metadata
 from unstructured.documents.html import HTMLDocument
 from unstructured.documents.xml import VALID_PARSERS
@@ -24,6 +25,7 @@ if TYPE_CHECKING:
 
 @process_metadata()
 @add_metadata_with_filetype(FileType.HTML)
+@add_chunking_strategy()
 def partition_html(
     filename: Optional[str] = None,
     file: Optional[IO[bytes]] = None,
@@ -39,6 +41,7 @@ def partition_html(
     metadata_filename: Optional[str] = None,
     metadata_last_modified: Optional[str] = None,
     skip_headers_and_footers: bool = False,
+    chunking_strategy: Optional[str] = None,
     **kwargs,
 ) -> List[Element]:
     """Partitions an HTML document into its constituent elements.