+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Enhancements
 
 * Updated HTML Partitioning to extract tables
+* Create and add `add_chunking_strategy` decorator to partition functions
 
 ### Features